spring boot jpa integration

+++ module-2-ms/code/springboot/productappv2/src/main/java/com/productapp/dao/ProductDaoImplUsingMap.java
@@ -13,8 +13,8 @@ public class ProductDaoImplUsingMap implements ProductDao{
 	
 	private static int counter;
 	static {
-		products.put(1, new Product(1, "Laptop", 500000));
-		products.put(2, new Product(2, "Laptop cover", 400));
+		products.put(1, new Product(1, "Laptop", BigDecimal.valueOf(500000)));
+		products.put(2, new Product(2, "Laptop cover", BigDecimal.valueOf(400)));
 		counter=2;
 	}
